Fix parameter overrides in inference.py
As a user I found it a bit more intuitive for a parameter set on a method to temporarily override a parameter set on class initialization. Setting the threshold on a method call and then not have it do anything would surprise me.

I may have misunderstood things so please feel free to just close this PR if I was under the wrong idea.